+++ jsonschema2pojo-core/src/test/java/org/jsonschema2pojo/util/InflectorTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright © 2010-2013 Nokia
+ * Copyright © 2010-2014 Nokia
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -48,5 +48,8 @@ public class InflectorTest {
         assertThat(Inflector.getInstance().singularize("sheep"), is("sheep"));
         assertThat(Inflector.getInstance().singularize("squid"), is("squid"));
         
+        assertThat(Inflector.getInstance().singularize("s"), is("s"));
+
     }
+    
 }